requirements.txt
requests
whois
argparse
beautifulsoup4
pillow
imagehash
urllib3
certifi
chardet
idna
lxml